Perfection

+++ 14-binary_tree_balance.c
@@ -8,15 +8,14 @@
 
 int binary_tree_balance(const binary_tree_t *tree)
 {
-	if (tree == NULL)
-	return (0);
+    if (tree == NULL)
+        return (0);
 
-	size_t left_height = binary_tree_height(tree->left);
-	size_t right_height = binary_tree_height(tree->right);
+    int left_height = (int)binary_tree_height(tree->left);
+    int right_height = (int)binary_tree_height(tree->right);
 
-	return ((int)(left_height - right_height));
+    return (left_height - right_height);
 }
-
 
 
 /**
